refine code structure
