+++ src/Raylib.jl
@@ -4,8 +4,12 @@ using Raylib_jll
 using CEnum
 
 include("color.jl")
+include("struct.jl")
 
-include("./binding/binding.jl")
+include("binding/binding.jl")
 using .Binding
+
+include("core.jl")
+include("shape.jl")
 
 end
