+++ src/shape.jl
@@ -1,0 +1,1 @@
+Binding.RayRectangle(p::RayVector2, width, height) = RayRectangle(p..., width, height)
